[TABP-33] Refactor booking update process and improve UpdateBookingAsync handling

+++ TABP/TABP.API/Helpers/EnumSchemaFilter.cs
@@ -20,7 +20,7 @@ namespace TABP.API.Helpers
             if (context.Type.IsEnum)
             {
                 var enumNames = Enum.GetNames(context.Type);
-                var enumValues = Enum.GetValues(context.Type).Cast<byte>();
+                var enumValues = Enum.GetValues(context.Type).Cast<int>();
                 var enumDescriptions = enumValues
                     .Zip(enumNames, (value, name) => $"{value} = {name}");
                 schema.Description += " Possible values: " + string.Join(", ", enumDescriptions);
